move ohd telemetry models into telemetry/models subdirectory
