+++ app/telemetry/models/wifiadapter.h
